refactor : HttpResponse 에 OutputStream write 역할 위임
